requirements.txt
rpi-rgb-led-matrix>=1.30.0
Pillow>=10.0.0
paho-mqtt>=1.6.0
PyYAML>=6.0
python-daemon>=2.3.0
